User avatar is an own component
Reason: Loading of user avatar contains logic
